fix: removed href from embedded entities
Technically, an embedded entity can be an embedded link or an embedded representation.
In the case of a link, the href is mandatory.  But in the case of a representation
it isn't even allowed (this is how you can tell which type you are dealing with).
But since the AddEntity function is handling the case of an embedded representation,
setting href was wrong.

Unfortunately even with generics Go doesn't provide a great way of handling these
union type cases.